user CRUD and middleware developed

+++ database/querys/manipulateTasks.js
@@ -11,7 +11,7 @@ async function createTable() {
     return new Promise((resolve, reject) => {
         const qry = `CREATE TABLE IF NOT EXISTS tasks (
             id INT PRIMARY KEY AUTO_INCREMENT,
-            user_id INT NOT NULL,
+            user_id VARCHAR(10) NOT NULL,
             name VARCHAR(50) NOT NULL,
             description VARCHAR(80),
             deadline DATE,
@@ -32,36 +32,55 @@ async function createTable() {
 
 // -- Manipulações no banco --
 
-export async function readTasks() {
+export async function readTasks(id) {
     /*
     readTasks() essa função retorna todas as instâncias de tarefas armazenadas no banco 
     */
     const con = await db; //inicializa o banco
     await createTable() //verifica se a tabela esta criada
 
-    const query = 'SELECT * FROM tasks' //query de busca
+
+
+    const query = 'SELECT * FROM tasks WHERE user_id = ?' //query de busca
 
     return new Promise((resolve, reject) => { //promise  com a resposta
-       con.query(query, (err, result) => { //função que realiza a busca em segundo plano da aplicação
+       con.query(query, [id], (err, result) => { //função que realiza a busca em segundo plano da aplicação
         err ? reject(err) : resolve(result);
        }) 
     });
 }
 
-export async function readTask(id) {
+export async function readTask(id, userId) {
     /*
     readTask(id) recebe o id de uma tarefa e retorna a sua instância no banco
     */
     const con = await db;
     await createTable();
 
-    const query = 'SELECT * FROM tasks WHERE id = ?';
-
-    return new Promise((resolve, reject) => {
-        con.query(query, [id], (err, result) => {
+    //resgata o id do proprietario da tarefa
+    const queryRead = 'SELECT * FROM tasks WHERE id = ?';
+    const taskData = await new Promise((resolve, reject) => {
+        con.query(queryRead, [id], (err, result) => {
             err ? reject(err) : resolve(result);
         })
     })
+    const taskUser = (taskData[0].user_id);
+
+    //verifica se o user da requisição é dono da tarefa
+    if (taskUser === userId){ 
+        const query = 'SELECT * FROM tasks WHERE id = ?';
+
+        return new Promise((resolve, reject) => {
+            con.query(query, [id], (err, result) => {
+                err ? reject(err) : resolve(result);
+            })
+        })
+    } else {
+        return Promise.reject({ 
+            status: 403, 
+            error: "Você não tem autorização para ler essa tarefa" 
+        });
+    }
 }
 
 export async function createTask(newTask) {
@@ -97,7 +116,7 @@ export async function createTask(newTask) {
     });
 }
 
-export async function updateTask(id, updatedTask) {
+export async function updateTask(id, updatedTask, userId) {
     /*
     updateTask(id, updatedTask) essa função atualiza os campos de uma tarefa. Ela recebe o id da tarefa alvo e um objeto com os novos valores. Caso a requsiição HTTP não tenha passado todos os campos, eles são definidos como null e o valor antigo permanece
     */
@@ -113,41 +132,65 @@ export async function updateTask(id, updatedTask) {
         })
     })
 
-    const queryUpdate = 'UPDATE tasks SET ? WHERE id = ?';
-    const values = [
-        { //objeto da tarefa
-            id: taskNow[0].id, //o id nunca pode ser editado
-            user_id: taskNow[0].user_id, //usuário nunca pode ser editado
-            name: updatedTask.name ?? taskNow[0].name,
-            description: updatedTask.description ?? taskNow[0].description,
-            deadline: updatedTask.deadline ?? taskNow[0].deadline,
-            urgency: updatedTask.urgency ?? taskNow[0].urgency,
-            status: updatedTask.status ?? taskNow[0].status,
-            created_at: taskNow[0].created_at //a data de criação não muda
-        }, 
-        id //id da tarefa
-    ]
+    //verifica se o usuário do request é dono da tarefa
+    const taskUser = (taskNow[0].user_id);
 
-    return new Promise((resolve, reject) => {
-        con.query(queryUpdate, values, (err, result) => {
-            err ? reject(err) : resolve(result);
+    if (taskUser === userId) {
+        const queryUpdate = 'UPDATE tasks SET ? WHERE id = ?';
+        const values = [
+            { //objeto da tarefa
+                id: taskNow[0].id, //o id nunca pode ser editado
+                user_id: taskNow[0].user_id, //usuário nunca pode ser editado
+                name: updatedTask.name ?? taskNow[0].name,
+                description: updatedTask.description ?? taskNow[0].description,
+                deadline: updatedTask.deadline ?? taskNow[0].deadline,
+                urgency: updatedTask.urgency ?? taskNow[0].urgency,
+                status: updatedTask.status ?? taskNow[0].status,
+                created_at: taskNow[0].created_at //a data de criação não muda
+            }, 
+            id //id da tarefa
+        ]
+
+        return new Promise((resolve, reject) => {
+            con.query(queryUpdate, values, (err, result) => {
+                err ? reject(err) : resolve(result);
+            })
         })
-    })
-
+    } else {
+        return Promise.reject({ 
+            status: 403, 
+            error: "Você não tem autorização para manipular essa tarefa" 
+        });
+    }
 }
 
-export async function deleteTask(id) {
+export async function deleteTask(id, userId) {
     /*
     deleteTask(id) recebe o id da tarefa e exclui a instância do banco
     */
     const con = await db; //inicializa o banco
     await createTable() //verifica se a tabela esta criada
 
-    const query = 'DELETE FROM tasks WHERE id = ?';
-
-    return new Promise((resolve, reject) => {
-        con.query(query, [id], (err, result) => {
+    const queryRead = 'SELECT * FROM tasks WHERE id = ?';
+    const taskData = await new Promise((resolve, reject) => {
+        con.query(queryRead, [id], (err, result) => {
             err ? reject(err) : resolve(result);
-        });
+        })
     })
+    const taskUser = (taskData[0].user_id);
+
+    if (taskUser === userId){
+        const query = 'DELETE FROM tasks WHERE id = ?';
+
+        return new Promise((resolve, reject) => {
+            con.query(query, [id], (err, result) => {
+                err ? reject(err) : resolve(result);
+            });
+        })
+    } else {
+        return Promise.reject({ 
+            status: 403, 
+            error: "Você não tem autorização para manipular essa tarefa" 
+        });
+    }
 }
